create new find-strategy.test file, FYI we could name it later differently
